redefine ino_t as int64_t

+++ kernel/include/stddef.h
@@ -17,7 +17,7 @@ typedef int64_t time_t;
 typedef int32_t pid_t;
 typedef unsigned mode_t;
 typedef uint64_t dev_t;
-typedef unsigned short ino_t;
+typedef uint64_t ino_t;
 typedef unsigned uid_t;
 typedef unsigned gid_t;
